feat: refactor stack service volume commands and schemas to use StackServiceVolumeTypeDto
- Updated CreateStackServiceVolumeCommand and UpdateStackServiceVolumeCommand to utilize StackServiceVolumeTypeDto for type consistency.
- Refactored validation classes to align with the new DTO structure, ensuring proper validation rules are applied.
- Modified related schemas and TypeScript types to reflect the changes in volume type handling.
- Added unit tests to verify the mapping between StackServiceVolumeType and StackServiceVolumeTypeDto, ensuring functionality remains intact.

+++ backend/tests/CloudCrafter.FunctionalTests/Domain/Stacks/StackServiceVolumes/GetStackServiceVolumesQueryTest.cs
@@ -66,7 +66,7 @@ public class GetStackServiceVolumesQueryTest : BaseStackServiceVolumeTest
     {
         dto.Id.Should().Be(volume.Id);
         dto.Name.Should().Be(volume.Name);
-        dto.Type.Should().Be(volume.Type);
+        ((int)dto.Type).Should().Be((int)volume.Type);
         dto.SourcePath.Should().Be(volume.SourcePath);
         dto.DestinationPath.Should().Be(volume.DestinationPath);
     }
